requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
itsdangerous==2.2.0
stripe==12.5.1
supabase==2.10.0
